IS-2720: Ktor 3.0
