+++ src/main/kotlin/no/nav/syfo/util/VeilederAPIAccessPipeline.kt
@@ -1,20 +1,19 @@
 package no.nav.syfo.util
 
-import io.ktor.server.application.*
-import io.ktor.util.pipeline.*
+import io.ktor.server.routing.*
 import no.nav.syfo.application.exception.ForbiddenAccessVeilederException
 import no.nav.syfo.client.veiledertilgang.VeilederTilgangskontrollClient
 import no.nav.syfo.domain.PersonIdentNumber
 
-suspend fun PipelineContext<out Unit, ApplicationCall>.validateVeilederAccess(
+suspend fun RoutingContext.validateVeilederAccess(
     action: String,
     personIdentToAccess: PersonIdentNumber,
     veilederTilgangskontrollClient: VeilederTilgangskontrollClient,
     requestBlock: suspend () -> Unit,
 ) {
-    val callId = getCallId()
+    val callId = this.call.getCallId()
 
-    val token = getBearerHeader()
+    val token = this.call.getBearerHeader()
         ?: throw IllegalArgumentException("Failed to complete the following action: $action. No Authorization header supplied")
 
     val hasVeilederAccess = veilederTilgangskontrollClient.hasAccess(
