+++ src/test/kotlin/no/nav/syfo/application/api/PodApiSpek.kt
@@ -1,8 +1,12 @@
+package no.nav.syfo.application.api
+
+import io.ktor.client.request.*
+import io.ktor.client.statement.*
 import io.ktor.http.*
 import io.ktor.server.routing.*
 import io.ktor.server.testing.*
 import no.nav.syfo.application.ApplicationState
-import no.nav.syfo.application.api.registerPodApi
+import no.nav.syfo.application.database.DatabaseInterface
 import no.nav.syfo.testhelper.TestDatabase
 import no.nav.syfo.testhelper.TestDatabaseNotResponding
 import org.amshove.kluent.shouldBeEqualTo
@@ -11,91 +15,90 @@ import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
 
 object PodApiSpek : Spek({
+    val database = TestDatabase()
+    val databaseNotResponding = TestDatabaseNotResponding()
 
-    describe("Successful liveness and readiness checks") {
-        with(TestApplicationEngine()) {
-            start()
-            val database = TestDatabase()
-            application.routing {
+    fun ApplicationTestBuilder.setupPodApi(database: DatabaseInterface, applicationState: ApplicationState) {
+        application {
+            routing {
                 registerPodApi(
-                    applicationState = ApplicationState(
-                        alive = true,
-                        ready = true
-                    ),
+                    applicationState = applicationState,
                     database = database,
                 )
             }
+        }
+    }
 
-            it("Returns ok on is_alive") {
-                with(handleRequest(HttpMethod.Get, "/internal/is_alive")) {
-                    response.status()?.isSuccess() shouldBeEqualTo true
-                    response.content shouldNotBeEqualTo null
-                }
+    describe("Successful liveness and readiness checks") {
+        it("Returns ok on is_alive") {
+            testApplication {
+                setupPodApi(
+                    database = database,
+                    applicationState = ApplicationState(alive = true, ready = true),
+                )
+                val response = client.get("/internal/is_alive")
+                response.status.isSuccess() shouldBeEqualTo true
+                response.bodyAsText() shouldNotBeEqualTo null
             }
-            it("Returns ok on is_alive") {
-                with(handleRequest(HttpMethod.Get, "/internal/is_ready")) {
-                    response.status()?.isSuccess() shouldBeEqualTo true
-                    response.content shouldNotBeEqualTo null
-                }
+        }
+        it("Returns ok on is_ready") {
+            testApplication {
+                setupPodApi(
+                    database = database,
+                    applicationState = ApplicationState(alive = true, ready = true),
+                )
+                val response = client.get("/internal/is_ready")
+                response.status.isSuccess() shouldBeEqualTo true
+                response.bodyAsText() shouldNotBeEqualTo null
             }
         }
     }
 
     describe("Unsuccessful liveness and readiness checks") {
-        with(TestApplicationEngine()) {
-            start()
-            val database = TestDatabase()
-            application.routing {
-                registerPodApi(
-                    ApplicationState(
-                        alive = false,
-                        ready = false
-                    ),
-                    database,
+        it("Returns internal server error when liveness check fails") {
+            testApplication {
+                setupPodApi(
+                    database = database,
+                    applicationState = ApplicationState(alive = false, ready = false),
                 )
+                val response = client.get("/internal/is_alive")
+                response.status shouldBeEqualTo HttpStatusCode.InternalServerError
+                response.bodyAsText() shouldNotBeEqualTo null
             }
-
-            it("Returns internal server error when liveness check fails") {
-                with(handleRequest(HttpMethod.Get, "/internal/is_alive")) {
-                    response.status() shouldBeEqualTo HttpStatusCode.InternalServerError
-                    response.content shouldNotBeEqualTo null
-                }
-            }
-
-            it("Returns internal server error when readiness check fails") {
-                with(handleRequest(HttpMethod.Get, "/internal/is_ready")) {
-                    response.status() shouldBeEqualTo HttpStatusCode.InternalServerError
-                    response.content shouldNotBeEqualTo null
-                }
+        }
+        it("Returns internal server error when readiness check fails") {
+            testApplication {
+                setupPodApi(
+                    database = database,
+                    applicationState = ApplicationState(alive = true, ready = false),
+                )
+                val response = client.get("/internal/is_ready")
+                response.status shouldBeEqualTo HttpStatusCode.InternalServerError
+                response.bodyAsText() shouldNotBeEqualTo null
             }
         }
     }
     describe("Successful liveness and unsuccessful readiness checks when database not working") {
-        with(TestApplicationEngine()) {
-            start()
-            val database = TestDatabaseNotResponding()
-            application.routing {
-                registerPodApi(
-                    ApplicationState(
-                        alive = true,
-                        ready = true
-                    ),
-                    database,
+        it("Returns internal server error when liveness check fails") {
+            testApplication {
+                setupPodApi(
+                    database = databaseNotResponding,
+                    applicationState = ApplicationState(alive = true, ready = false),
                 )
+                val response = client.get("/internal/is_alive")
+                response.status.isSuccess() shouldBeEqualTo true
+                response.bodyAsText() shouldNotBeEqualTo null
             }
-
-            it("Returns ok on is_alive") {
-                with(handleRequest(HttpMethod.Get, "/internal/is_alive")) {
-                    response.status()?.isSuccess() shouldBeEqualTo true
-                    response.content shouldNotBeEqualTo null
-                }
-            }
-
-            it("Returns internal server error when readiness check fails") {
-                with(handleRequest(HttpMethod.Get, "/internal/is_ready")) {
-                    response.status() shouldBeEqualTo HttpStatusCode.InternalServerError
-                    response.content shouldNotBeEqualTo null
-                }
+        }
+        it("Returns internal server error when readiness check fails") {
+            testApplication {
+                setupPodApi(
+                    database = databaseNotResponding,
+                    applicationState = ApplicationState(alive = true, ready = true),
+                )
+                val response = client.get("/internal/is_ready")
+                response.status shouldBeEqualTo HttpStatusCode.InternalServerError
+                response.bodyAsText() shouldNotBeEqualTo null
             }
         }
     }
