+++ src/test/kotlin/no/nav/syfo/unntak/api/UnntakApiSpek.kt
@@ -2,7 +2,12 @@ package no.nav.syfo.unntak.api
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
+import io.ktor.client.*
+import io.ktor.client.plugins.contentnegotiation.*
+import io.ktor.client.request.*
+import io.ktor.client.statement.*
 import io.ktor.http.*
+import io.ktor.serialization.jackson.*
 import io.ktor.server.testing.*
 import io.mockk.*
 import no.nav.syfo.dialogmotekandidat.database.getDialogmotekandidatEndringListForPerson
@@ -30,237 +35,248 @@ class UnntakApiSpek : Spek({
     val urlUnntakStatistikk = "$unntakApiBasePath/$unntakApiStatistikk"
 
     describe(UnntakApiSpek::class.java.simpleName) {
-        with(TestApplicationEngine()) {
-            start()
-            val externalMockEnvironment = ExternalMockEnvironment.instance
-            val database = externalMockEnvironment.database
-            val kafkaProducer = mockk<KafkaProducer<String, KafkaDialogmotekandidatEndring>>()
-            val dialogmotekandidatEndringProducer = DialogmotekandidatEndringProducer(
-                kafkaProducerDialogmotekandidatEndring = kafkaProducer,
-            )
+        val externalMockEnvironment = ExternalMockEnvironment.instance
+        val database = externalMockEnvironment.database
+        val kafkaProducer = mockk<KafkaProducer<String, KafkaDialogmotekandidatEndring>>()
+        val dialogmotekandidatEndringProducer = DialogmotekandidatEndringProducer(
+            kafkaProducerDialogmotekandidatEndring = kafkaProducer,
+        )
 
-            application.testApiModule(
-                externalMockEnvironment = externalMockEnvironment,
-                dialogmotekandidatEndringProducer = dialogmotekandidatEndringProducer,
-            )
-
-            beforeEachTest {
-                database.dropData()
-
-                clearMocks(kafkaProducer)
-                coEvery {
-                    kafkaProducer.send(any())
-                } returns mockk<Future<RecordMetadata>>(relaxed = true)
+        fun ApplicationTestBuilder.setupApiAndClient(): HttpClient {
+            application {
+                testApiModule(
+                    externalMockEnvironment = externalMockEnvironment,
+                    dialogmotekandidatEndringProducer = dialogmotekandidatEndringProducer,
+                )
             }
+            val client = createClient {
+                install(ContentNegotiation) {
+                    jackson { configure() }
+                }
+            }
+            return client
+        }
 
-            val validToken = generateJWT(
-                audience = externalMockEnvironment.environment.azure.appClientId,
-                issuer = externalMockEnvironment.wellKnownInternalAzureAD.issuer,
-                navIdent = UserConstants.VEILEDER_IDENT,
-            )
-            val newUnntakDTO = generateNewUnntakDTO(personIdent = UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER)
+        beforeEachTest {
+            database.dropData()
 
-            describe("Get unntak for person") {
-                describe("Happy path") {
-                    it("returns unntak for person if request is successful") {
+            clearMocks(kafkaProducer)
+            coEvery {
+                kafkaProducer.send(any())
+            } returns mockk<Future<RecordMetadata>>(relaxed = true)
+        }
+
+        val validToken = generateJWT(
+            audience = externalMockEnvironment.environment.azure.appClientId,
+            issuer = externalMockEnvironment.wellKnownInternalAzureAD.issuer,
+            navIdent = UserConstants.VEILEDER_IDENT,
+        )
+        val newUnntakDTO = generateNewUnntakDTO(personIdent = UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER)
+
+        describe("Get unntak for person") {
+            describe("Happy path") {
+                it("returns unntak for person if request is successful") {
+                    testApplication {
+                        val client = setupApiAndClient()
+
                         val unntak = newUnntakDTO.toUnntak(createdByIdent = UserConstants.VEILEDER_IDENT)
                         database.connection.use {
                             it.createUnntak(unntak)
                             it.commit()
                         }
 
-                        with(
-                            handleRequest(HttpMethod.Get, urlUnntakPersonIdent) {
-                                addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
-                                addHeader(NAV_PERSONIDENT_HEADER, UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER.value)
-                            }
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.OK
+                        val response = client.get(urlUnntakPersonIdent) {
+                            bearerAuth(validToken)
+                            header(NAV_PERSONIDENT_HEADER, UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER.value)
+                        }
+                        response.status shouldBeEqualTo HttpStatusCode.OK
 
-                            val unntakList = objectMapper.readValue<List<UnntakDTO>>(response.content!!)
-                            unntakList.size shouldBeEqualTo 1
+                        val unntakList = objectMapper.readValue<List<UnntakDTO>>(response.bodyAsText())
+                        unntakList.size shouldBeEqualTo 1
 
-                            val unntakDTO = unntakList.first()
-                            unntakDTO.createdAt shouldNotBeEqualTo null
-                            unntakDTO.uuid shouldNotBeEqualTo null
-                            unntakDTO.personIdent shouldBeEqualTo UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER.value
-                            unntakDTO.createdBy shouldBeEqualTo UserConstants.VEILEDER_IDENT
-                            unntakDTO.arsak shouldBeEqualTo newUnntakDTO.arsak
-                            unntakDTO.beskrivelse shouldBeEqualTo newUnntakDTO.beskrivelse
-                        }
-                    }
-                }
-                describe("Unhappy paths") {
-                    it("returns status Unauthorized if no token is supplied") {
-                        with(
-                            handleRequest(HttpMethod.Get, urlUnntakPersonIdent) {}
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.Unauthorized
-                        }
-                    }
-                    it("returns status Forbidden if denied access to person") {
-                        with(
-                            handleRequest(HttpMethod.Get, urlUnntakPersonIdent) {
-                                addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
-                                addHeader(
-                                    NAV_PERSONIDENT_HEADER,
-                                    UserConstants.PERSONIDENTNUMBER_VEILEDER_NO_ACCESS.value
-                                )
-                            }
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.Forbidden
-                        }
-                    }
-                    it("should return status BadRequest if no $NAV_PERSONIDENT_HEADER is supplied") {
-                        with(
-                            handleRequest(HttpMethod.Get, urlUnntakPersonIdent) {
-                                addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
-                            }
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.BadRequest
-                        }
-                    }
-                    it("should return status BadRequest if $NAV_PERSONIDENT_HEADER with invalid PersonIdent is supplied") {
-                        with(
-                            handleRequest(HttpMethod.Get, urlUnntakPersonIdent) {
-                                addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
-                                addHeader(
-                                    NAV_PERSONIDENT_HEADER,
-                                    UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER.value.drop(1)
-                                )
-                            }
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.BadRequest
-                        }
+                        val unntakDTO = unntakList.first()
+                        unntakDTO.createdAt shouldNotBeEqualTo null
+                        unntakDTO.uuid shouldNotBeEqualTo null
+                        unntakDTO.personIdent shouldBeEqualTo UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER.value
+                        unntakDTO.createdBy shouldBeEqualTo UserConstants.VEILEDER_IDENT
+                        unntakDTO.arsak shouldBeEqualTo newUnntakDTO.arsak
+                        unntakDTO.beskrivelse shouldBeEqualTo newUnntakDTO.beskrivelse
                     }
                 }
             }
+            describe("Unhappy paths") {
+                it("returns status Unauthorized if no token is supplied") {
+                    testApplication {
+                        val client = setupApiAndClient()
+                        val response = client.get(urlUnntakPersonIdent) {}
+                        response.status shouldBeEqualTo HttpStatusCode.Unauthorized
+                    }
+                }
+                it("returns status Forbidden if denied access to person") {
+                    testApplication {
+                        val client = setupApiAndClient()
 
-            describe("Create unntak for person") {
-                describe("Happy path") {
-                    it("creates Unntak and DialogmotekandidatEndring (not kandidat) when person is kandidat") {
+                        val response = client.get(urlUnntakPersonIdent) {
+                            bearerAuth(validToken)
+                            header(NAV_PERSONIDENT_HEADER, UserConstants.PERSONIDENTNUMBER_VEILEDER_NO_ACCESS.value)
+                        }
+                        response.status shouldBeEqualTo HttpStatusCode.Forbidden
+                    }
+                }
+                it("should return status BadRequest if no $NAV_PERSONIDENT_HEADER is supplied") {
+                    testApplication {
+                        val client = setupApiAndClient()
+
+                        val response = client.get(urlUnntakPersonIdent) {
+                            bearerAuth(validToken)
+                        }
+                        response.status shouldBeEqualTo HttpStatusCode.BadRequest
+                    }
+                }
+                it("should return status BadRequest if $NAV_PERSONIDENT_HEADER with invalid PersonIdent is supplied") {
+                    testApplication {
+                        val client = setupApiAndClient()
+
+                        val response = client.get(urlUnntakPersonIdent) {
+                            bearerAuth(validToken)
+                            header(NAV_PERSONIDENT_HEADER, UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER.value.drop(1))
+                        }
+                        response.status shouldBeEqualTo HttpStatusCode.BadRequest
+                    }
+                }
+            }
+        }
+
+        describe("Create unntak for person") {
+            describe("Happy path") {
+                it("creates Unntak and DialogmotekandidatEndring (not kandidat) when person is kandidat") {
+                    testApplication {
+                        val client = setupApiAndClient()
                         val dialogmotekandidatEndring = generateDialogmotekandidatEndringStoppunkt(
                             personIdentNumber = UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER,
                         )
                         database.createDialogmotekandidatEndring(dialogmotekandidatEndring = dialogmotekandidatEndring)
-                        with(
-                            handleRequest(HttpMethod.Post, urlUnntakPersonIdent) {
-                                addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
-                                addHeader(HttpHeaders.ContentType, ContentType.Application.Json.toString())
-                                setBody(objectMapper.writeValueAsString(newUnntakDTO))
-                            }
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.Created
-                            val producerRecordSlot = slot<ProducerRecord<String, KafkaDialogmotekandidatEndring>>()
-                            verify(exactly = 1) {
-                                kafkaProducer.send(capture(producerRecordSlot))
-                            }
+                        val response = client.post(urlUnntakPersonIdent) {
+                            bearerAuth(validToken)
+                            header(HttpHeaders.ContentType, ContentType.Application.Json.toString())
+                            setBody(objectMapper.writeValueAsString(newUnntakDTO))
+                        }
 
-                            val latestDialogmotekandidatEndring =
-                                database.connection.getDialogmotekandidatEndringListForPerson(
-                                    personIdent = UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER
-                                ).firstOrNull()
-                            latestDialogmotekandidatEndring shouldNotBeEqualTo null
-                            latestDialogmotekandidatEndring!!.kandidat shouldBeEqualTo false
-                            latestDialogmotekandidatEndring.arsak shouldBeEqualTo DialogmotekandidatEndringArsak.UNNTAK.name
+                        response.status shouldBeEqualTo HttpStatusCode.Created
+                        val producerRecordSlot = slot<ProducerRecord<String, KafkaDialogmotekandidatEndring>>()
+                        verify(exactly = 1) {
+                            kafkaProducer.send(capture(producerRecordSlot))
+                        }
 
-                            val latestUnntak =
-                                database.getUnntakList(
-                                    personIdent = UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER
-                                ).first()
-                            latestUnntak.createdAt shouldNotBeEqualTo null
-                            latestUnntak.personIdent shouldBeEqualTo UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER.value
-                            latestUnntak.createdBy shouldBeEqualTo UserConstants.VEILEDER_IDENT
-                            latestUnntak.arsak shouldBeEqualTo newUnntakDTO.arsak
-                            latestUnntak.beskrivelse shouldBeEqualTo newUnntakDTO.beskrivelse
+                        val latestDialogmotekandidatEndring =
+                            database.connection.getDialogmotekandidatEndringListForPerson(
+                                personIdent = UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER
+                            ).firstOrNull()
+                        latestDialogmotekandidatEndring shouldNotBeEqualTo null
+                        latestDialogmotekandidatEndring!!.kandidat shouldBeEqualTo false
+                        latestDialogmotekandidatEndring.arsak shouldBeEqualTo DialogmotekandidatEndringArsak.UNNTAK.name
 
-                            val kafkaDialogmoteKandidatEndring = producerRecordSlot.captured.value()
-                            kafkaDialogmoteKandidatEndring.personIdentNumber shouldBeEqualTo UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER.value
-                            kafkaDialogmoteKandidatEndring.arsak shouldBeEqualTo DialogmotekandidatEndringArsak.UNNTAK.name
-                            kafkaDialogmoteKandidatEndring.kandidat shouldBeEqualTo false
-                            kafkaDialogmoteKandidatEndring.unntakArsak shouldBeEqualTo newUnntakDTO.arsak
+                        val latestUnntak =
+                            database.getUnntakList(
+                                personIdent = UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER
+                            ).first()
+                        latestUnntak.createdAt shouldNotBeEqualTo null
+                        latestUnntak.personIdent shouldBeEqualTo UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER.value
+                        latestUnntak.createdBy shouldBeEqualTo UserConstants.VEILEDER_IDENT
+                        latestUnntak.arsak shouldBeEqualTo newUnntakDTO.arsak
+                        latestUnntak.beskrivelse shouldBeEqualTo newUnntakDTO.beskrivelse
+
+                        val kafkaDialogmoteKandidatEndring = producerRecordSlot.captured.value()
+                        kafkaDialogmoteKandidatEndring.personIdentNumber shouldBeEqualTo UserConstants.ARBEIDSTAKER_PERSONIDENTNUMBER.value
+                        kafkaDialogmoteKandidatEndring.arsak shouldBeEqualTo DialogmotekandidatEndringArsak.UNNTAK.name
+                        kafkaDialogmoteKandidatEndring.kandidat shouldBeEqualTo false
+                        kafkaDialogmoteKandidatEndring.unntakArsak shouldBeEqualTo newUnntakDTO.arsak
+                    }
+                }
+            }
+            describe("Unhappy paths") {
+                it("returns status Unauthorized if no token is supplied") {
+                    testApplication {
+                        val client = setupApiAndClient()
+                        val response = client.post(urlUnntakPersonIdent) {
+                            header(HttpHeaders.ContentType, ContentType.Application.Json.toString())
+                            setBody(objectMapper.writeValueAsString(newUnntakDTO))
+                        }
+                        response.status shouldBeEqualTo HttpStatusCode.Unauthorized
+                        verify(exactly = 0) {
+                            kafkaProducer.send(any())
                         }
                     }
                 }
-                describe("Unhappy paths") {
-                    it("returns status Unauthorized if no token is supplied") {
-                        with(
-                            handleRequest(HttpMethod.Post, urlUnntakPersonIdent) {}
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.Unauthorized
-                            verify(exactly = 0) {
-                                kafkaProducer.send(any())
-                            }
-                        }
-                    }
-                    it("returns status Forbidden if denied access to person") {
+                it("returns status Forbidden if denied access to person") {
+                    testApplication {
+                        val client = setupApiAndClient()
                         val newUnntakDTOWithDeniedAccess =
                             generateNewUnntakDTO(personIdent = UserConstants.PERSONIDENTNUMBER_VEILEDER_NO_ACCESS)
-                        with(
-                            handleRequest(HttpMethod.Post, urlUnntakPersonIdent) {
-                                addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
-                                addHeader(HttpHeaders.ContentType, ContentType.Application.Json.toString())
-                                setBody(objectMapper.writeValueAsString(newUnntakDTOWithDeniedAccess))
-                            }
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.Forbidden
-                            verify(exactly = 0) {
-                                kafkaProducer.send(any())
-                            }
+                        val response = client.post(urlUnntakPersonIdent) {
+                            bearerAuth(validToken)
+                            header(HttpHeaders.ContentType, ContentType.Application.Json.toString())
+                            setBody(objectMapper.writeValueAsString(newUnntakDTOWithDeniedAccess))
+                        }
+                        response.status shouldBeEqualTo HttpStatusCode.Forbidden
+                        verify(exactly = 0) {
+                            kafkaProducer.send(any())
                         }
                     }
-                    it("returns Conflict when person is not kandidat") {
-                        with(
-                            handleRequest(HttpMethod.Post, urlUnntakPersonIdent) {
-                                addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
-                                addHeader(HttpHeaders.ContentType, ContentType.Application.Json.toString())
-                                setBody(objectMapper.writeValueAsString(newUnntakDTO))
-                            }
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.Conflict
-                            verify(exactly = 0) {
-                                kafkaProducer.send(any())
-                            }
+                }
+                it("returns Conflict when person is not kandidat") {
+                    testApplication {
+                        val client = setupApiAndClient()
+                        val response = client.post(urlUnntakPersonIdent) {
+                            bearerAuth(validToken)
+                            header(HttpHeaders.ContentType, ContentType.Application.Json.toString())
+                            setBody(objectMapper.writeValueAsString(newUnntakDTO))
+                        }
+                        response.status shouldBeEqualTo HttpStatusCode.Conflict
+                        verify(exactly = 0) {
+                            kafkaProducer.send(any())
                         }
                     }
                 }
             }
-            describe("Get unntaksstatistikk for veileder") {
-                describe("Happy path") {
-                    it("returns unntaksstatistikk if request is successful") {
+        }
+        describe("Get unntaksstatistikk for veileder") {
+            describe("Happy path") {
+                it("returns unntaksstatistikk if request is successful") {
+                    testApplication {
+                        val client = setupApiAndClient()
                         val unntak = newUnntakDTO.toUnntak(createdByIdent = UserConstants.VEILEDER_IDENT)
                         database.connection.use {
                             it.createUnntak(unntak)
                             it.commit()
                         }
-
-                        with(
-                            handleRequest(HttpMethod.Get, urlUnntakStatistikk) {
-                                addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
-                            }
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.OK
-
-                            val unntakStatistikkList =
-                                objectMapper.readValue<List<UnntakStatistikk>>(response.content!!)
-                            unntakStatistikkList.size shouldBeEqualTo 1
-
-                            val unntakStatistikk = unntakStatistikkList.first()
-                            unntakStatistikk.unntakDato shouldNotBeEqualTo null
-                            unntakStatistikk.tilfelleStart shouldNotBeEqualTo null
-                            unntakStatistikk.tilfelleEnd shouldNotBeEqualTo null
+                        val response = client.get(urlUnntakStatistikk) {
+                            bearerAuth(validToken)
                         }
+                        response.status shouldBeEqualTo HttpStatusCode.OK
+
+                        val unntakStatistikkList =
+                            objectMapper.readValue<List<UnntakStatistikk>>(response.bodyAsText())
+                        unntakStatistikkList.size shouldBeEqualTo 1
+
+                        val unntakStatistikk = unntakStatistikkList.first()
+                        unntakStatistikk.unntakDato shouldNotBeEqualTo null
+                        unntakStatistikk.tilfelleStart shouldNotBeEqualTo null
+                        unntakStatistikk.tilfelleEnd shouldNotBeEqualTo null
                     }
                 }
-                describe("Unhappy paths") {
-                    it("returns status Unauthorized if no token is supplied") {
-                        with(
-                            handleRequest(HttpMethod.Get, urlUnntakPersonIdent) {}
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.Unauthorized
-                        }
+            }
+            describe("Unhappy paths") {
+                it("returns status Unauthorized if no token is supplied") {
+                    testApplication {
+                        val client = setupApiAndClient()
+                        val response = client.get(urlUnntakStatistikk) {}
+                        response.status shouldBeEqualTo HttpStatusCode.Unauthorized
                     }
-                    it("returns empty unntaksstatistikk if no access to person with unntak") {
+                }
+                it("returns empty unntaksstatistikk if no access to person with unntak") {
+                    testApplication {
+                        val client = setupApiAndClient()
                         val unntak =
                             generateNewUnntakDTO(personIdent = UserConstants.PERSONIDENTNUMBER_VEILEDER_NO_ACCESS).toUnntak(
                                 createdByIdent = UserConstants.VEILEDER_IDENT
@@ -269,18 +285,14 @@ class UnntakApiSpek : Spek({
                             it.createUnntak(unntak)
                             it.commit()
                         }
-
-                        with(
-                            handleRequest(HttpMethod.Get, urlUnntakStatistikk) {
-                                addHeader(HttpHeaders.Authorization, bearerHeader(validToken))
-                            }
-                        ) {
-                            response.status() shouldBeEqualTo HttpStatusCode.OK
-
-                            val unntakStatistikkList =
-                                objectMapper.readValue<List<UnntakStatistikk>>(response.content!!)
-                            unntakStatistikkList.size shouldBeEqualTo 0
+                        val response = client.get(urlUnntakStatistikk) {
+                            bearerAuth(validToken)
                         }
+                        response.status shouldBeEqualTo HttpStatusCode.OK
+
+                        val unntakStatistikkList =
+                            objectMapper.readValue<List<UnntakStatistikk>>(response.bodyAsText())
+                        unntakStatistikkList.size shouldBeEqualTo 0
                     }
                 }
             }
